refactor: step3 PR반영
- 띄어수기 수정
- Mode 추가, 내부에서 mode를 변경하도록 수정
- 이미지 aspectfill 로 변경
- extension 분리
