+++ OpenMarket/OpenMarket/View/ProductListCell.xib
@@ -17,7 +17,7 @@
                 <stackView opaque="NO" contentMode="scaleToFill" alignment="center" spacing="10" translatesAutoresizingMaskIntoConstraints="NO" id="j5F-mp-G3I">
                     <rect key="frame" x="10" y="10" width="121.5" height="130"/>
                     <subviews>
-                        <imageView clipsSubviews="YES" userInteractionEnabled="NO" contentMode="scaleAspectFit" horizontalHuggingPriority="251" verticalHuggingPriority="251" translatesAutoresizingMaskIntoConstraints="NO" id="YFL-wQ-nVl">
+                        <imageView clipsSubviews="YES" userInteractionEnabled="NO" contentMode="scaleAspectFill" horizontalHuggingPriority="251" verticalHuggingPriority="251" translatesAutoresizingMaskIntoConstraints="NO" id="YFL-wQ-nVl">
                             <rect key="frame" x="0.0" y="30" width="70" height="70"/>
                             <color key="backgroundColor" systemColor="labelColor"/>
                             <constraints>
