+++ OpenMarket/OpenMarket/ViewController/ProductEditViewController.swift
@@ -93,17 +93,69 @@ class ProductEditViewController: UIViewController {
     }
         
     // MARK: Helpers
-    private func registerKeyboardNotificationObserver(){
+    func setMode(mode: Mode) {
+        switch mode {
+        case .add:
+            self.mode = AddMode(self)
+        case .edit:
+            self.mode = EditMode(self)
+        }
+    }
+    
+    func setProductInfo() {
+        productNameTextField.text = product?.name
+        productPriceTextField.text = product?.price.description
+        productStockTextField.text = product?.stock.description
+        productCurrencySegment.selectedSegmentIndex = product?.currency == .usd ? 1 : 0
+        discountedPriceTextField.text = product?.discountedPrice.description
+        productDescriptionTextView.text = product?.description
+        
+        loadImages()
+    }
+
+    enum Mode {
+        case add, edit
+    }
+}
+
+private extension ProductEditViewController {
+    func loadImages() {
+        guard let product = product, let images = product.images else { return }
+        self.images = [UIImage?](repeating: nil, count: images.count)
+        for (idx, image) in images.enumerated() {
+            guard let imageURL = image.imageUrl else { return }
+            ImageCache.shared.load(url: imageURL as NSURL) { productImage in
+                DispatchQueue.main.async {
+                    self.images[idx] = productImage
+                }
+            }
+        }
+    }
+    
+    func setupCollectionView() {
+        imageCollectionView.delegate = self
+        imageCollectionView.dataSource = self
+        imageCollectionView.collectionViewLayout = flowlayout
+    }
+    
+    func addTapGesture() {
+        let tapGesture = UITapGestureRecognizer(target: self, action: #selector(UIInputViewController.dismissKeyboard))
+        tapGesture.cancelsTouchesInView = false
+        tapGesture.delegate = self
+        view.addGestureRecognizer(tapGesture)
+    }
+    
+    func registerKeyboardNotificationObserver() {
         NotificationCenter.default.addObserver(self, selector: #selector(keyboardWillShow), name: UIResponder.keyboardWillShowNotification, object: nil)
         NotificationCenter.default.addObserver(self, selector: #selector(keyboardWillClose), name: UIResponder.keyboardWillHideNotification, object: nil)
     }
     
-    private func unregisterKeyboardNotificationObserver(){
+    func unregisterKeyboardNotificationObserver() {
         NotificationCenter.default.removeObserver(self, name: UIResponder.keyboardWillShowNotification, object: nil)
         NotificationCenter.default.removeObserver(self, name: UIResponder.keyboardWillHideNotification, object: nil)
     }
     
-    private func validateProductInput() -> ProductInput?{
+    func validateProductInput() -> ProductInput?{
         guard let name = productNameTextField.text,
               let descriptions = productDescriptionTextView.text,
               let priceString = productPriceTextField.text else { return nil }
@@ -127,87 +179,5 @@ class ProductEditViewController: UIViewController {
         let currency: Currency = productCurrencySegment.selectedSegmentIndex == 1 ? .usd : .krw
         
         return ProductInput(name: name, descriptions: descriptions, price: price, currency: currency, discountedPrice: discountedPrice, stock: stock)
-    }
-    
-    func setProductInfo() {
-        productNameTextField.text = product?.name
-        productPriceTextField.text = product?.price.description
-        productStockTextField.text = product?.stock.description
-        productCurrencySegment.selectedSegmentIndex = product?.currency == .usd ? 1 : 0
-        discountedPriceTextField.text = product?.discountedPrice.description
-        productDescriptionTextView.text = product?.description
-        
-        loadImages()
-    }
-    
-    private func setupCollectionView(){
-        imageCollectionView.delegate = self
-        imageCollectionView.dataSource = self
-        imageCollectionView.collectionViewLayout = flowlayout
-    }
-    
-    private func addTapGesture(){
-        let tapGesture = UITapGestureRecognizer(target: self, action: #selector(UIInputViewController.dismissKeyboard))
-        tapGesture.cancelsTouchesInView = false
-        tapGesture.delegate = self
-        view.addGestureRecognizer(tapGesture)
-    }
-    
-    private func loadImages() {
-        guard let product = product, let images = product.images else { return }
-        self.images = [UIImage?](repeating: nil, count: images.count)
-        for (idx, image) in images.enumerated() {
-            guard let imageURL = image.imageUrl else { return }
-            ImageCache.shared.load(url: imageURL as NSURL) { productImage in
-                DispatchQueue.main.async {
-                    self.images[idx] = productImage
-                }
-            }
-        }
-    }
-}
-
-extension ProductEditViewController: UICollectionViewDataSource, UICollectionViewDelegate {
-    func collectionView(_ collectionView: UICollectionView, numberOfItemsInSection section: Int) -> Int {
-        return mode?.imageCount() ?? 0
-    }
-    
-    func collectionView(_ collectionView: UICollectionView, cellForItemAt indexPath: IndexPath) -> UICollectionViewCell {
-        if indexPath.row == images.count && images.count < 5, let _ = mode as? AddMode {
-            let cell = imageCollectionView.dequeueReusableCell(withReuseIdentifier: "LastImageCell", for: indexPath)
-            return cell
-        } else {
-            let cell = imageCollectionView.dequeueReusableCell(withReuseIdentifier: "ProductImageCell", for: indexPath) as! ProductImageCell
-            cell.setImage(images[indexPath.row])
-            return cell
-        }
-    }
-    
-    func collectionView(_ collectionView: UICollectionView, didSelectItemAt indexPath: IndexPath) {
-        mode?.didSelectItem(index: indexPath.row)
-    }
-}
-
-extension ProductEditViewController: UIImagePickerControllerDelegate, UINavigationControllerDelegate {
-    func imagePickerController(_ picker: UIImagePickerController, didFinishPickingMediaWithInfo info: [UIImagePickerController.InfoKey : Any]) {
-        guard let newImage = info[UIImagePickerController.InfoKey.editedImage] as? UIImage,
-              let selectedImageIdx = selectedImageIdx
-            else {
-                picker.dismiss(animated: true, completion: nil)
-                return
-            }
-        if images.count > selectedImageIdx {
-            images[selectedImageIdx] = newImage
-        } else {
-            images.append(newImage)
-        }
-        picker.dismiss(animated: true, completion: nil)
-    }
-}
-
-extension ProductEditViewController: UIGestureRecognizerDelegate {
-    func gestureRecognizer(_ gestureRecognizer: UIGestureRecognizer, shouldReceive touch: UITouch) -> Bool {
-        let touchPoint = touch.location(in: view)
-        return !imageCollectionView.frame.contains(touchPoint)
     }
 }
